Fix Comb allocator bugs: consolidate New/Delete, add BuddyAllocator::Reset
Consolidate duplicate comb::New/Delete from allocator_concepts.h and new.h
into a single concept-constrained implementation with trivially_destructible
optimization. Fix misleading ThreadSafeAllocator::Underlying() comment, fix
incorrect peakBytesUsed aggregation in GlobalMemoryTracker, add missing
BuddyAllocator::Reset(), and add missing #include <cstdio> in platform_utils.h.
