+++ Comb/tests/test_pool.cpp
@@ -1,5 +1,6 @@
 #include <larvae/larvae.h>
 #include <comb/pool_allocator.h>
+#include <comb/new.h>
 #include <cstring>
 
 namespace
